requirements.txt
numpy==1.13.3
scikit-learn==0.19.0
# scipy==0.19.1
Theano==0.7.0

Pillow==6.2.2